vuln-scanner/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
termcolor>=2.3.0

# API clients
shodan>=1.31.0
censys>=2.2.0

# Utilities
python-nmap>=0.7.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
Jinja2>=3.1.2

# Output and reporting
rich>=13.6.0
tabulate>=0.9.0
pandas>=2.1.0

# Async and concurrency
aiohttp>=3.9.0
asyncio>=3.4.3

# Progress and CLI
tqdm>=4.66.0
click>=8.1.7
pyfiglet>=1.0.2

# Data validation
validators>=0.22.0
dnspython>=2.4.2
